rename: movieController > movieListApp으로 변경
